requirements.txt
pyyaml==5.*